requirements.clean.txt
requests>=2.31.0
python-dotenv>=1.0.0
memoripy>=0.1.2
ollama>=0.3.3
beautifulsoup4>=4.12.3
lxml>=5.2.0
mcp[cli]>=1.6.0
